+++ app/src/main/java/com/jifalops/localization/datatypes/TofRanging.java
@@ -18,7 +18,7 @@ public class TofRanging {
     public TofRanging(String[] csv) {
         id1 = csv[0];
         id2 = csv[1];
-        tof = Integer.valueOf(csv[2]);
+        tof = Float.valueOf(csv[2]);
         dist = Float.valueOf(csv[3]);
         range = Float.valueOf(csv[4]);
         fspl = Float.valueOf(csv[5]);
